bugfix upload image

+++ model/members/index.js
@@ -6,14 +6,14 @@ const randtoken = require('rand-token');
 const bcrypt = require('bcrypt');
 const multer = require("multer");
 const fs = require("fs");
+var path = require('path');
 const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, 'public')
-  },
-  filename: function (req, file, cb) {
-    cb(null, file.fieldname + Date.now() + '.jpg');
+  destination : path.join(__dirname + './../../public'),
+  filename: function(req, file, cb){
+      cb(null, file.fieldname + Date.now() +
+      path.extname(file.originalname));
   }
-})
+});
 
 exports.index = (req, res) => {
   connection.query("SELECT * FROM members", (error, payload) => {
@@ -82,34 +82,35 @@ exports.updatePassword = (req, res) => {
 }
 
 exports.updateImage = (req, res) => {
-
   const upload = multer({ storage : storage}).single("image");
-    upload(req,res,function(err) {
-      const checkImage = connection.query(`SELECT image FROM members WHERE id = '${req.body.id}'`, (err, result) => {
-        err ? response.err({ code: err.code }, err) : result
-        const oldImage = result[0].image;
-          if(oldImage != null){
-            fs.unlink("./public/"+oldImage, (err) => {
-              if (err) {
-                response.err({ code: err.code }, err);
-              }
-            });
-          }
+
+    upload(req,res,function(error) {
+
+      if(!req.file){
+        response.err({ message: "invalid data request" }, res)
+      } else {
+        connection.query(`SELECT image FROM members WHERE id = '${req.body.id}'`, (err, result) => {
+          const oldImage = result[0].image;
+            if(oldImage != null){
+              
+              fs.unlink("./public/"+oldImage, (err) => {
+                if (err) {
+                  response.err({ code: err.code }, res);
+                }
+              });
+            }
         });
-
-      if(checkImage){
-
-        if(err) {
+        
+        if(error) {
           return response.err({ message: "upload error" }, res);
         } else {
           connection.query(`UPDATE members SET image = '${req.file.filename}' where id = '${req.body.id}'`, (error, payload) => {
-            error ? response.err({ code: error.code }, error) : response.ok({ data: payload.affectedRows }, res)
+            error ? response.err({ code: error.code }, res) : response.ok({ data: payload.affectedRows }, res)
           });
         }
       }
-      
     });
-}
+  }
 
 
 exports.login = (req, res) => {
